+++ src/main/java/info/freelibrary/sodbox/impl/WeakHashTable.java
@@ -8,7 +8,7 @@ public class WeakHashTable implements OidHashTable {
     static final float loadFactor = 0.75f;
     int count;
     int threshold;
-    int nModified;
+    long nModified;
     boolean disableRehash;
     StorageImpl db;
 
@@ -91,7 +91,7 @@ public class WeakHashTable implements OidHashTable {
         while (true) { 
             cs:synchronized(this) { 
                 disableRehash = true;
-                int n;
+                long n;
                 do { 
                     n = nModified;
                     for (int i = 0; i < table.length; i++) {
